chat_with_data/requirements.txt
streamlit
langchain
langchain-openai
pandas
openai
plotly
python-dotenv
matplotlib
tabulate